строчу, как пулемет. с хакатона столько времени не писал. добавил удаление, не добавил диалог на подтверждение. еще и маркеры надо. но фотографий сервер больше не принимает. пишет, что размер не тот, хотя до 2Мб.

+++ app/src/main/java/ed/maevski/testbalinasoft/data/cache/entity/ImageEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "IMAGES")
 data class ImageEntity(
-    @field: ColumnInfo(name = "id") @field:PrimaryKey(autoGenerate = true) var id: Long = 0,
+    @field: ColumnInfo(name = "id") @field:PrimaryKey(autoGenerate = true) var id: Int = 0,
     @field: ColumnInfo(name = "uri") val uri: String? = null,
     @field: ColumnInfo(name = "url") val url: String? = null,
     @field: ColumnInfo(name = "date") val date: Long? = null,
